Code running and works but need to fix some bugs

+++ database/models/user.py
@@ -2,7 +2,6 @@ from database.base import Base, int64, str_40
 from sqlalchemy.orm import Mapped
 
 from typing import Optional
-from decimal import Decimal
 
 from sqlalchemy.orm import mapped_column
 
@@ -12,7 +11,7 @@ class User(Base):
 
     user_id: Mapped[int64] = mapped_column(primary_key=True)
     username: Mapped[Optional[str_40]] = mapped_column(unique=True)
-    sens_level: Mapped[Decimal] = mapped_column(default=0.5)
+    sens_level: Mapped[float] = mapped_column(default=0.5)
     data_folder: Mapped[str]
     result_folder: Mapped[str]
